websockets/constructor/014.html

<!doctype html>
<title>WebSockets: serialize establish a connection</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="../constants.js?pipe=sub"></script>
<div id=log></div>
<script>
var ws = [];
var events = 0;
var prevDate;
var date;
for (var i = 0; i < 4; ++i) {
  ws[i] = new WebSocket(SCHEME_DOMAIN_PORT+'/handshake_sleep_1');
  ws[i].id = i;
  ws[i].onopen = function(e) {
    events++;
    date = new Date();
    if (prevDate) {
      //Should really be > 1000, but we allow for 32ms differences from timer resolution
      assert_greater_than(date-prevDate, 968);
    }
    prevDate = date;
    this.onopen = function() {assert_unreached()};
  }
  ws[i].onclose = function(e) {
    events++;
    if (events == 8) {
      done();
    }
    this.onclose = function() {assert_unreached()};
  };
  ws[i].onerror = ws[i].onmessage = function() {assert_unreached()};
}
</script>
